Prepare nav for new games
Add an optional contextual navigation to the header.

Additionally, move some controllers to their respective context.
