+++ src/ConnectFour/Port/Adapter/Http/View/game.html.twig
@@ -1,4 +1,4 @@
-{% extends 'layout/condensed.html.twig' %}
+{% extends '@connect-four/base.html.twig' %}
 {% from "macro/username.html.twig" import render as renderUsername %}
 
 {% set page_title = game.redPlayerId and game.yellowPlayerId
