+++ src/ConnectFour/Port/Adapter/Http/View/home-tile.html.twig
@@ -1,0 +1,25 @@
+<div class="card">
+    <div class="card-body d-flex flex-column">
+        <h3 class="card-title">Connect Four</h3>
+        <div class="gp-cf-game gp-cf-game--disabled mb-3">
+            <div class="gp-cf-game__field"></div>
+            <div class="gp-cf-game__field"></div>
+            <div class="gp-cf-game__field"><div class="gp-cf-token gp-cf-token--yellow"></div></div>
+            <div class="gp-cf-game__field"><div class="gp-cf-token gp-cf-token--yellow"></div></div>
+            <div class="gp-cf-game__field"></div>
+            <div class="gp-cf-game__field"></div>
+            <div class="gp-cf-game__field"></div>
+            <div class="gp-cf-game__field"></div>
+            <div class="gp-cf-game__field"><div class="gp-cf-token gp-cf-token--yellow"></div></div>
+            <div class="gp-cf-game__field gp-cf-game__field--highlight"><div class="gp-cf-token"></div></div>
+            <div class="gp-cf-game__field gp-cf-game__field--highlight"><div class="gp-cf-token"></div></div>
+            <div class="gp-cf-game__field gp-cf-game__field--highlight"><div class="gp-cf-token"></div></div>
+            <div class="gp-cf-game__field gp-cf-game__field--highlight"><div class="gp-cf-token"></div></div>
+            <div class="gp-cf-game__field"></div>
+        </div>
+        <p class="small">
+            Drop tokens, line them up, and outsmart your opponent.
+        </p>
+        <a href="{{ path('connect_four_lobby') }}" class="btn btn-outline-primary">Play</a>
+    </div>
+</div>
