+++ src/ConnectFour/Port/Adapter/Http/View/lobby.html.twig
@@ -1,4 +1,4 @@
-{% extends 'layout/condensed.html.twig' %}
+{% extends '@connect-four/base.html.twig' %}
 
 {% set page_title = 'Lobby' %}
