+++ templates/bundles/TwigBundle/Exception/error404.html.twig
@@ -21,7 +21,7 @@
                 We are sorry, but the page you are looking for was not found.
             </p>
             <p class="text-center">
-                <a href="{{ path('lobby') }}" class="btn btn-primary">
+                <a href="{{ path('home') }}" class="btn btn-primary">
                     <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none"
                          stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"
                          class="icon">
