+++ templates/include/pagination.html.twig
@@ -1,23 +1,43 @@
-{% set totalPages = (total / limit)|round(0, 'ceil') %}
-{% set previousPage = min(max(page - 1, 1), totalPages) %}
-{% set nextPage = max(min(page + 1, totalPages), 2) %}
-<div class="card">
-    <div class="card-body">
-        <ul class="pagination">
-            <li class="page-item page-prev{{ page <= 1 ? ' disabled' }}">
-                <a class="page-link"
-                   href="{{ path('profile', app.request.query.all|merge({'page': previousPage})) }}">
-                    <div class="page-item-subtitle">previous</div>
-                    <div class="page-item-title">{{ previousTitle }}</div>
+{# Shoutout to https://gist.github.com/maxpou/612359ed4af4cc5c4f06 for the original code. #}
+
+{% set totalPages = (totalItems / itemsPerPage)|round(0, 'ceil') %}
+{% set visiblePages = visiblePages|default(3) %}
+{% set queryParameter = queryParameter|default('page') %}
+
+<ul class="pagination pagination-outline">
+    {% for i in 1..totalPages %}
+        {% if 0 == (currentPage - visiblePages) - loop.index %}
+            <li class="page-item">
+                <a href="{{ (app.request.pathInfo ~ '?' ~ (app.request.query.all|merge({(queryParameter): 1})|url_encode))|e }}"
+                   class="page-link">
+                    1
                 </a>
             </li>
-            <li class="page-item page-next{{ page >= totalPages ? ' disabled' }}">
-                <a class="page-link"
-                   href="{{ path('profile', app.request.query.all|merge({'page': nextPage})) }}">
-                    <div class="page-item-subtitle">next</div>
-                    <div class="page-item-title">{{ nextTitle }}</div>
+            {% if 1 != loop.index %}
+                <li class="page-item">
+                    <span class="page-link disabled">&hellip;</span>
+                </li>
+            {% endif %}
+        {% elseif 0 == (currentPage + visiblePages) - loop.index  and (currentPage + visiblePages) < totalPages %}
+            <li class="page-item">
+                <span class="page-link disabled">&hellip;</span>
+            </li>
+        {% elseif 0 < (currentPage - visiblePages) - loop.index %}
+        {% elseif 0 > (currentPage + visiblePages) - loop.index %}
+        {% else %}
+            <li class="page-item{{ currentPage == loop.index ? ' active' }}">
+                <a href="{{ (app.request.pathInfo ~ '?' ~ (app.request.query.all|merge({(queryParameter): loop.index})|url_encode))|e }}"
+                   class="page-link">
+                    {{ loop.index }}
                 </a>
             </li>
-        </ul>
-    </div>
-</div>
+        {% endif %}
+    {% endfor %}
+    {% if currentPage != totalPages and (currentPage + visiblePages) < totalPages %}
+        <li class="page-item">
+            <a href="{{ (app.request.pathInfo ~ '?' ~ (app.request.query.all|merge({(queryParameter): totalPages})|url_encode))|e }}"
+               class="page-link">{{ totalPages }}
+            </a>
+        </li>
+    {% endif %}
+</ul>
