+++ templates/layout/center-tight.html.twig
@@ -5,7 +5,7 @@
     <div class="page page-center" data-page-content>
         <div class="container container-tight">
             <p class="text-center fw-medium my-5">
-                <a href="{{ path('lobby') }}" class="link-secondary navbar-brand-autodark text-decoration-none">
+                <a href="{{ path('home') }}" class="link-secondary navbar-brand-autodark text-decoration-none">
                     <img src="{{ asset('img/logo.png') }}"
                          alt="Gaming Platform"
                          width="32"
