+++ templates/layout/partial/footer-links.html.twig
@@ -1,5 +1,5 @@
 <li class="list-inline-item">
-    <a href="{{ path('lobby') }}"
+    <a href="{{ path('home') }}"
        target="_blank"
        class="link-secondary"
        rel="noopener">
@@ -7,7 +7,7 @@
     </a>
 </li>
 <li class="list-inline-item">
-    <a href="{{ path('lobby') }}"
+    <a href="{{ path('home') }}"
        target="_blank"
        class="link-secondary"
        rel="noopener">
@@ -15,7 +15,7 @@
     </a>
 </li>
 <li class="list-inline-item">
-    <a href="{{ path('lobby') }}"
+    <a href="{{ path('home') }}"
        target="_blank"
        class="link-secondary"
        rel="noopener">
